health/pyproject.toml
[project]
name = "health"
version = "0.1.0"
description = "Swakopmund Health Service"
authors = [
    {name = "Sakaria Nghivafe",email = "sakarianghivafe@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-oauth-toolkit (>=3.0.1,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "daphne (>=4.1.2,<5.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)"
]

[tool.poetry]

[tool.poetry.scripts]
serve = "scripts.dev:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
